registros eliminar y cambiar fechas
seeder para accion masiva

+++ app/Http/Controllers/EmpleadosAusentismosController.php
@@ -306,15 +306,25 @@ class EmpleadosAusentismosController extends Controller
 	{
 		$ausentismo = Ausentismo::find($id);
 
-		if ($ausentismo->archivo != null) {
-		$ruta = storage_path("app/ausentismos/trabajador/{$ausentismo->id}");
-		$ruta_archivo = storage_path("app/ausentismos/trabajador/{$ausentismo->id}/{$ausentismo->hash_archivo}");
-		unlink($ruta_archivo);
-		rmdir($ruta);
+		if ($ausentismo) {
+			// Eliminar el archivo si existe
+			$ruta_archivo = storage_path("app/ausentismos/trabajador/{$ausentismo->id}/{$ausentismo->hash_archivo}");
+			file_exists($ruta_archivo) && unlink($ruta_archivo);
+	
+			// Eliminar la carpeta si existe
+			$ruta_carpeta = storage_path("app/ausentismos/trabajador/{$ausentismo->id}");
+			file_exists($ruta_carpeta) && rmdir($ruta_carpeta);
+	
+			// Eliminar las comunicaciones asociadas al ausentismo
+			Comunicacion::where('id_ausentismo', $id)->delete();
+	
+			// Finalmente, eliminar el ausentismo
+			$ausentismo->delete();
+			return back()->with('success', 'Ausentismo y Comunicación asociada eliminados correctamente');
+		}else{
+			return back()->with('error', 'El ausentismo no existe. Consulte a soporte si tiene dudas');
 		}
-		$comunicacion = Comunicacion::where('id_ausentismo', $id)->delete();
-		$ausentismo->delete();
-		return back()->with('success', 'Ausentismo y Comunicación asociada eliminados correctamente');
+
 	}
